ai-code/requirements.txt
# Core ML Libraries
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# MLflow & Azure ML
mlflow>=2.9.0
azureml-core>=1.56.0
azureml-mlflow>=1.56.0
azure-ai-ml>=1.12.0
azure-identity>=1.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
pylint>=3.0.0

# Utilities
python-dotenv>=1.0.0
